pyproject.toml
[tool.poetry]
name = "nzoth"
version = "0.1.0"
description = "Easy way to deploy ML models"
authors = ["Eric Venarusso"]
license = "Apache2"

[tool.poetry.dependencies]
python = "^3.9"
cookiecutter = "^1.7.3"
PyYAML = "^6.0"
pydantic = "^1.9.0"
click = "^8.0.4"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
nzoth = "nzoth.cli:create"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
